+++ service-routing/src/main/java/fi/nls/oskari/routing/RoutingService.java
@@ -1,9 +1,11 @@
 package fi.nls.oskari.routing;
 
+import java.util.List;
+
 /**
  * Created by SMAKINEN on 26.6.2015.
  */
 public interface RoutingService {
 
-    Route getRouteBetween(double lon, double lat);
+    List<RouteResponse> getRoute(RouteParams params);
 }